avivo-mini-rag-bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1

faiss-cpu==1.7.4
sentence-transformers==2.6.1

transformers==4.41.2
huggingface-hub==0.23.4

numpy==1.26.4
requests==2.31.0

tqdm==4.66.4
pillow==10.3.0
